Criteria

RankChoiceVoting.CondorcetLoser — Type

CondorcetLoser <: Condorcet

A Condorcet loser criterion object. The Condorcet criterion states that a candiate who wins all pairwise elections must also win the election using a given voting system.

RankChoiceVoting.count_violations — Method

count_violations(system::VotingSystem, criterion::CondorcetLoser, rankings::Ranks; _...)

Counts the number of violations of the Condorcet loser criterion for a given voting system.

Arguments

system::VotingSystem: a voting system object

criterion::CondorcetLoser: condorcet criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

RankChoiceVoting.satisfies — Method

satisfies(system::VotingSystem, criterion::CondorcetLoser, rankings::Ranks; _...)

Tests whether a voting system satisfies the Condorcet loser criterion.

Arguments

system::VotingSystem: a voting system object

criterion::CondorcetLoser: condorcet loser criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

RankChoiceVoting.CondorcetWinner — Type

CondorcetWinner <: Criterion

A Condorcet winner criterion object. The Condorcet criterion states that a candiate who wins all pairwise elections must also win the election using a given voting system.

RankChoiceVoting.count_violations — Method

count_violations(system::VotingSystem, criterion::CondorcetWinner, rankings::Ranks; _...)

Counts the number of violations of the Condorcet for a given voting system.

Arguments

system::VotingSystem: a voting system object

criterion::CondorcetWinner: condorcet criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

RankChoiceVoting.satisfies — Method

satisfies(system::VotingSystem, criterion::CondorcetWinner, rankings::Ranks; _...)

Tests whether a voting system satisfies the Condorcet winner criterion.

Arguments

system::VotingSystem: a voting system object

criterion::CondorcetWinner: condorcet criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

RankChoiceVoting.Consistency — Type

Consistency <: Criterion

A Consistency criterion object. According to the consistency criterion, if the votes are split into disjoint subsets, and the same candidate wins each subset, the system must select the same winner for the whole set of votes.

RankChoiceVoting.count_violations — Method

count_violations(system::VotingSystem, criterion::Consistency, rankings::Ranks; n_reps=1000, _...)

Counts the number of violations of the consistency criterion for a given voting system.

Arguments

system::VotingSystem: a voting system object

criterion::Consistency: condorcet criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

Keywords

n_reps: number of Monte Carlo simulations to perform

RankChoiceVoting.satisfies — Method

satisfies(system::VotingSystem, criterion::Consistency, rankings::Ranks; n_max=1000, _...)

Tests whether a voting system satisfies the Consistency criterion.

Arguments

system::VotingSystem: a voting system object

criterion::Consistency: consistency criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

Keywords

n_reps: maximum Monte Carlo simulations to perform

RankChoiceVoting.Independence — Type

Independence <: Criterion

An object representing the independence of irrelevant alternatives criterion

RankChoiceVoting.count_violations — Method

count_violations(system::VotingSystem, criterion::Independence, rankings::Ranks; _...)

Counts the number of violations of the Independence of Irrelevant alternatives criterion. There are several ways to test this criterion. Currently, it is tested by removing subsets of losing candidates and testing whether the winner changes. Other methods could add and/or subtract candidates and test whether the rank order changes. The current method is less strict compared to alternative methods.

Arguments

system::VotingSystem: a voting system object

criterion::Independence: Independence criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

RankChoiceVoting.satisfies — Method

satisfies(::Fails, system::VotingSystem, criterion::Independence, rankings::Ranks;  _...)

Tests whether a voting system satisfies the Independence of Irrelevant alternatives criterion. There are several ways to test this criterion. Currently, it is tested by removing subsets of losing candidates and testing whether the winner changes. Other methods could add and/or subtract candidates and test whether the rank order changes. The current method is less strict compared to alternative methods.

Arguments

system::VotingSystem: a voting system object

criterion::Independence: Independence criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

RankChoiceVoting.Majority — Type

Majority <: Criterion

An object for the fairness criterion, majority. A voting system satisfies the majority criterion if it elects the candidate who recieves more than 50% first preferences.

RankChoiceVoting.count_violations — Method

count_violations(system::VotingSystem, criterion::Majority, rankings::Ranks; _...)

Counts the number of violations of the majority criterion for a given voting system.

Arguments

system::VotingSystem: a voting system object

criterion::Majority: majority criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

RankChoiceVoting.satisfies — Method

satisfies(system::VotingSystem, criterion::Majority, rankings::Ranks; _...)

Tests whether a voting system satisfies the majority criterion.

Arguments

system::VotingSystem: a voting system object

criterion::Majority: majority criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

RankChoiceVoting.Monotonicity — Type

Monotonicity <: Criterion

An object for the fairness criterion monotonicity. A voting system that satisfies the monotonicity criterion cannot elect a different candidate by increasing votes (without otherwise changing the rank order) of a candidate who would otherwise win. In other words, redistributing more votes to the winner (without otherwise changing the rank order), should not prevent the winner from winning.

RankChoiceVoting.count_violations — Method

count_violations(system::VotingSystem, criteria::Monotonicity, rankings::Ranks; n_reps=1000, _...)

Counts the number of violations of the monotonicity criterion for a given voting system.

Arguments

system::VotingSystem: a voting system object

criteria::Monotonicity: condorcet criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

Keywords

n_reps=1000: maximum number of Monte Carlo simulations to perform

RankChoiceVoting.satisfies — Method

satisfies(system::VotingSystem, criteria::Monotonicity, rankings::Ranks; _...)

Tests whether a voting system satisfies the monotonicity criterion.

Arguments

system::VotingSystem: a voting system object

criteria::Monotonicity: monotonicity criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

Keywords

max_reps=1000: maximum number of Monte Carlo simulations to perform while searching

for a violation

RankChoiceVoting.ReversalSymmetry — Type

ReversalSymmetry <: Criterion

An object for the fairness criterion reversal symmetry. According to the reversal systemetry criterion, a winner of an election cannot win if each voter's rankings are reversed

RankChoiceVoting.count_violations — Method

count_violations(system::VotingSystem, criterion::ReversalSymmetry, rankings::Ranks; _...)

Counts the number of violations of the reversal symmetry criterion for a given voting system.

Arguments

system::VotingSystem: a voting system object

criterion::ReversalSymmetry: reversal symmetry criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

RankChoiceVoting.satisfies — Method

satisfies(system::VotingSystem, criterion::ReversalSymmetry, rankings::Ranks; _...)

Tests whether a voting system satisfies the reversal symmetry criterion.

Arguments

system::VotingSystem: a voting system object

criterion::ReversalSymmetry: condorcet criterion object

rankings::Ranks: a rank choice voting object consisting of rank counts and unique ranks

Systems

RankChoiceVoting.Borda — Type

Borda <: VotingSystem

A Borda count voting system object.

RankChoiceVoting.compute_ranks — Method

compute_ranks(system::Borda, rankings::Ranks)

Ranks candidates using Borda count system.

Arguments

system: a Borda voting system object

rankings::Ranks: an object containing counts and unique rank orders

RankChoiceVoting.evaluate_winner — Method

evaluate_winner(system::Borda, rankings::Ranks)

Returns the id of the winning candiate in Borda count system.

Arguments

system: a Borda voting system object

rankings::Ranks: an object containing counts and unique rank orders

RankChoiceVoting.Bucklin — Type

Bucklin{T,I<:Integer} <: VotingSystem{T,I}

A Bucklin voting system object.

RankChoiceVoting.compute_ranks — Method

compute_ranks(system::Bucklin, rankings::Ranks)

Ranks candidates using the Bucklin system.

Arguments

system: a Borda voting system object

rankings::Ranks: an object containing counts and unique rank orders

RankChoiceVoting.evaluate_winner — Method

evaluate_winner(system::Bucklin, rankings::Ranks)

Returns the id of the winning candiate using the Bucklin voting system.

Arguments

system: a Bucklin voting system object

RankChoiceVoting.InstantRunOff — Type

InstantRunOff{T,I<:Integer} <: VotingSystem{T,I}

An instant runoff voting system object.

Arguments

uranks: a vector of unique rankings. Each ranking is a vector in which index represents rank and value represents candidate id.

counts: a vector of frequency counts corresponding to each unique ranking

RankChoiceVoting.compute_ranks — Method

compute_ranks(system::InstantRunOff, rankings::Ranks)

Ranks candidates using the InstantRunOff system.

Arguments

system: an InstantRunOff voting system object

rankings::Ranks: an object containing counts and unique rank orders

RankChoiceVoting.evaluate_winner — Method

evaluate_winner(system::InstantRunOff, rankings::Ranks)

Returns the id of the winning candiate in an instant runoff election.

Arguments

system: an instant runoff system object

rankings::Ranks: an object containing counts and unique rank orders